update coordinated assault
